Add collide_plot() and to_data()
`collide_plot` returns True if the coordinates lie within the plot area, and False otherwise. This is useful when testing if a touch lies within the plot area (for example, if the user is supposed to interact with the plot).

`to_data` converts window coordinates into data coordinates, using the Graph's xmin, xmax, ymin, and ymax, a la `to_local` et al.